Moved helper logic into a utils file. Cleaner and easier to reuse
